Fix 6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int b;
 long int c;
 long long int d;
 float f;
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of an integer: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char: %d byte(s)\n", (int)sizeof(a));
+printf("Size of an integer: %d byte(s)\n", (int)sizeof(b));
+printf("Size of a long int: %d byte(s)\n", (int)sizeof(c));
+printf("Size of a long long int: %d byte(s)\n", (int)sizeof(d));
+printf("Size of a float: %d byte(s)\n", (int)sizeof(f));
 return (0);
 }
